Change functionality so that operations are handled using an associated dictionary
The event listeners have been modified so that they depend on an "operation" function that handles the events and then delegates operations to distinct operation functions.

The purpose of this approach is to make the code a bit more modular and generalisable, in the event that the interface is extended to include more complex functionality.